+++ HISTORY.rst
@@ -16,6 +16,8 @@ Unreleased
 
 **Bugfixes**
 
+* Fix a bug with orderpoint on product template
+
 **Build**
 
 **Documentation**